+++ sdk/storage/azure_storage_blob/src/models/mod.rs
@@ -21,6 +21,7 @@ pub use download_result::{
 pub use method_options::BlobClientDownloadOptions;
 pub use method_options::BlockBlobClientUploadOptions;
 pub use method_options::BlockBlobClientUploadOptions as BlobClientUploadOptions;
+pub use method_options::{BlobContainerClientListBlobsOptions, ListBlobsAcceptFormat};
 pub use upload_result::BlockBlobClientUploadResult;
 pub use upload_result::BlockBlobClientUploadResult as BlobClientUploadResult;